Restructure app: * Modular structure separated between dashboard, auth, calendar, insights, and workout modules. * Added flask blueprints.
