+++ ouraapp/auth/helpers.py
@@ -5,9 +5,11 @@ import time
 from oura import OuraClient
 from dotenv import load_dotenv
 from flask_login import current_user
-from ouraapp import database
-from ouraapp import db
-from sqlalchemy import desc, func
+from ouraapp.models import Day
+from .models import User
+from ouraapp.dashboard.models import Sleep, Readiness
+from ouraapp.calendar.models import Events
+from ouraapp.extensions import db
 import logging
 
 logger = logging.getLogger("ouraapp")
@@ -24,29 +26,6 @@ def pull_oura_data():
     sleep_json = json.dumps(sleep_summary)
     readiness_json = json.dumps(readiness_summary)
     return [sleep_json, readiness_json]
-
-
-def update_days_db():
-    last_id = db.session.query(func.max(database.Day.id)).scalar()
-    if last_id:
-        next_id = last_id + 1
-        last_day_obj = database.Day.query.filter(
-            database.Day.id == last_id).first()
-        start_date = last_day_obj.date + timedelta(days=1)
-    else:
-        next_id = 0
-        start_date = date(2022, 1, 1)
-    added_days = [
-        start_date + timedelta(days=x)
-        for x in range((date.today() - start_date).days + 1)
-    ]
-    for i, day in enumerate(added_days, start=next_id):
-        day_obj = database.Day()
-        day_obj.id = i
-        day_obj.date = day
-        day_obj.date_str = day.strftime('%Y-%m-%d')
-        db.session.add(day_obj)
-    db.session.commit()
 
 
 def date_hook(json_dict):
@@ -80,7 +59,7 @@ def convert_seconds(total_seconds):
 
 def add_event_to_db(new_event_dict):
     json_event = json.dumps(new_event_dict)
-    event = database.Events(event=json_event, user_id=current_user.id)
+    event = Events(event=json_event, user_id=current_user.id)
     db.session.add(event)
     db.session.commit()
 
@@ -89,23 +68,23 @@ def add_sleep_to_db(json_dict):
     ''' Commit data to sleep database model.'''
     selected_data = json.loads(json_dict, object_hook=date_hook)
     logger.debug(
-        f'User is authenticated: user obj = {database.User.query.filter_by(id=current_user.id).first()}'
+        f'User is authenticated: user obj = {User.query.filter_by(id=current_user.id).first()}'
     )
     for entry in (selected_data['sleep']):
         day = format_date(entry['summary_date'])
-        db_day = database.Day.query.filter_by(date=day).first()
+        db_day = Day.query.filter_by(date=day).first()
         logger.debug(
-            f'{db_day.id}: - 1) Both user_id and sleep data are in database: {database.Sleep.query.filter_by(user_id = current_user.id, id=db_day.id).first()}'
+            f'{db_day.id}: - 1) Both user_id and sleep data are in database: {Sleep.query.filter_by(user_id = current_user.id, id=db_day.id).first()}'
         )
         logger.debug(
-            f'{db_day.id}: - 2) user_id in database: {database.Sleep.query.filter_by(id=db_day.id).first()}'
+            f'{db_day.id}: - 2) user_id in database: {Sleep.query.filter_by(id=db_day.id).first()}'
         )
-        if db.session.query(database.Sleep).filter_by(
+        if db.session.query(Sleep).filter_by(
                 id=db_day.id, user_id=current_user.id).count() < 1:
             logger.debug(
                 f'{db_day.id}: 3) Query says sleep obj should be added to db. {db_day.date}. Associated id: {db_day.id}.'
             )
-            prev_night_data = database.Sleep(
+            prev_night_data = Sleep(
                 date=day,
                 id=db_day.id,
                 sleep_score=entry['score'],
@@ -137,10 +116,10 @@ def add_readiness_to_db(json_dict):
     selected_data = json.loads(json_dict, object_hook=date_hook)
     for entry in (selected_data['readiness']):
         day = format_date(entry['summary_date'])
-        db_day = database.Day.query.filter_by(date=day).first()
-        if db.session.query(database.Readiness).filter_by(
+        db_day = Day.query.filter_by(date=day).first()
+        if db.session.query(Readiness).filter_by(
                 date=day, user_id=current_user.id).count() < 1:
-            prev_night_data = database.Readiness(
+            prev_night_data = Readiness(
                 date=day,
                 id=db_day.id,
                 hrv_balance=entry['score_hrv_balance'],
