+++ ouraapp/calendar/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('calendar', __name__)
+
+from ouraapp.calendar import routes
