+++ ouraapp/insights/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('insights', __name__)
+
+from ouraapp.insights import routes
