+++ ouraapp/templates/base.html
@@ -7,7 +7,7 @@
             <a class="navbar-brand" href="#">Jack's Wellness App</a>
             <div class="navbar-nav">
                 <li class="nav-item">
-                    <a class="nav-link" href="/log">Wellness Log</a>
+                    <a class="nav-link" href="/dashboard/log">Wellness Log</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="/calendar">Calendar</a>
@@ -16,10 +16,10 @@
                     <a class="nav-link" href="/insights">Insights</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/register">Register</a>
+                    <a class="nav-link" href="/auth/register">Register</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/logout">Log Out</a>
+                    <a class="nav-link" href="/auth/logout">Log Out</a>
                 </li>
             </div>
         </div>
@@ -27,7 +27,7 @@
     </nav>
 
 
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='main.css') }}" />
+    <!-- <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='main.css') }}" /> -->
 
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
