+++ ouraapp/templates/index.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% block head %}
-
+<!-- <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='main.css') }}" /> -->
 {% endblock %}
 
 {% block content %}
@@ -41,7 +41,7 @@
 
         <div>
             <br />
-            <span>Don't have an account? <a href="{{ url_for('login') }}">Sign up.</a></span></br>
+            <span>Don't have an account? <a href="{{ url_for('auth.register') }}">Sign up.</a></span></br>
         </div>
     </form>
