+++ ouraapp/weights/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('weights', __name__)
+
+from ouraapp.weights import routes
