+++ run.py
@@ -1,8 +1,6 @@
 from ouraapp import create_app
 import logging
 
-
-
 app = create_app()
 
-from ouraapp import routes
+# from ouraapp import routes
